requirements.txt
requests~=2.31.0
gradio~=4.29.0
python-dotenv~=1.0.1